pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "group-member-guardian-bot"
version = "0.1.0"
description = "PTB bot scaffold for tracking group members and removing users from all groups"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=22.0,<23.0",
  "SQLAlchemy>=2.0,<3.0",
  "aiosqlite>=0.20,<1.0",
  "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
  "ruff>=0.5,<1.0",
  "mypy>=1.10,<2.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
